Extract pool/download functions from catalog_openalex into openalex_pool
catalog_openalex.py was 659 lines (above the 500-line smell threshold).
The download, budget, sidecar, and filter functions form a cohesive unit
that other scripts may reuse, so extracting them into openalex_pool.py
is the natural seam. build_filter moved with the download functions
because fetch_query and _download_tiers depend on it directly -- keeping
it in catalog_openalex would require either circular imports or a callback
injection pattern.

catalog_openalex.py re-exports all moved names so existing imports from
other test files (test_retry_unification, test_collect_config) keep working.
test_openalex_from_date imports updated to openalex_pool since it
monkey-patches LAST_RUN_PATH (must patch the defining module).

Closes #620.
